+++ lib/screens/animal_reg_found/animal_reg_found.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 //Acompanhe desing do projeto aqui --> https://www.figma.com/file/GYFrt79mzIbOUXXmFyDgwL/Material-Baseline-Design-Kit?node-id=38%3A5814
 
-class cadastroAnimalEncontrado extends StatelessWidget {
+class AnimalRegisterFound extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
